backend/lambda-python/requirements-dev.txt
pytest>=8.0.0
pytest-cov>=5.0.0
moto[s3,dynamodb]>=5.0.0
ruff>=0.6.0
mypy>=1.11.0
boto3-stubs[bedrock-agent-runtime,kendra,dynamodb,s3]>=1.35.0
